requirements.txt
absl-py==2.1.0
accelerate==0.34.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.0a10
docker-pycreds==0.4.0
einops==0.8.0
etils==1.9.4
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
grpcio==1.66.1
gymnasium==1.0.0
huggingface-hub==0.25.0
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
mujoco==3.2.3
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
nvitop==1.3.2
pandas==2.2.2
pillow==10.2.0
protobuf==5.28.1
pycparser==2.22
PyOpenGL==3.1.7
pyparsing==3.1.4
pyrallis==0.3.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.14.0
setproctitle==1.3.3
smmap==5.0.1
sympy==1.12
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchsummary==1.5.1
torchvision==0.19.1+cu124
tqdm==4.66.5
transformers==4.40.0
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.3
wandb==0.18.2
Werkzeug==3.0.4